My first commit

+++ 0x02-functions_nested_loops/7-print_last_digit.c
@@ -6,7 +6,7 @@
  * Return: the last digit
  */
 
-int print_last_digit(int n)
+int print_last_digit(long int n)
 {
 	if (n < 0)
 	{
